+++ packages/scripts/config/webpack.config.js
@@ -5,7 +5,14 @@ const { BundleAnalyzerPlugin } = require( 'webpack-bundle-analyzer' );
 const { CleanWebpackPlugin } = require( 'clean-webpack-plugin' );
 const CopyWebpackPlugin = require( 'copy-webpack-plugin' );
 const browserslist = require( 'browserslist' );
-const MiniCSSExtractPlugin = require( 'mini-css-extract-plugin' );
+
+/*
+ * Temporary workaround to fix broken CommonJS import in version 2.5.0.
+ * @see https://github.com/WordPress/gutenberg/issues/37992
+ */
+const MiniCSSExtractPlugin = require( 'mini-css-extract-plugin' ).default
+	? require( 'mini-css-extract-plugin' ).default
+	: require( 'mini-css-extract-plugin' );
 const { basename, dirname, resolve } = require( 'path' );
 const ReactRefreshWebpackPlugin = require( '@pmmmwh/react-refresh-webpack-plugin' );
 const TerserPlugin = require( 'terser-webpack-plugin' );